Move input class
